Rewrite of msatCommand.  Entire replacement of original file.
Given some inherent problems in the initial way that msatCommand was 
coded (and, by extension, msatcommander), I basically decided that 
an entire rewrite of the code was both expedient and useful.

Largely, this means that many of the original functions have been moved
to classes within seqsearch.py.  msatCommand.py, itself, has been rewritten
to reflect these changes, making it shorter (and slightly faster).  Additionally,
I added tests for the classes along with a much more complete set of test inputs 
in testRepeats.fa.

Some functions remain incomplete, but these should be fixed in the next day or 2.